+++ react/src/data/map_relationship.json
@@ -1,1 +1,1 @@
-[["State", "County"], ["Native Area", "Native Subdivision"], ["Native Statistical Area", "Native Subdivision"], ["CSA", "MSA"], ["MSA", "County"], ["County", "City"], ["CCD", "City"], ["City", "Neighborhood"], ["School District", "Neighborhood"], ["ZIP", "Neighborhood"], ["Urban Area", "City"], ["Judicial Circuit", "Judicial District"], ["County", "County"], ["MSA", "MSA"], ["CSA", "CSA"], ["Urban Area", "Urban Area"], ["ZIP", "ZIP"], ["CCD", "CCD"], ["City", "City"], ["Neighborhood", "Neighborhood"], ["Congressional District", "Congressional District"], ["State House District", "State House District"], ["State Senate District", "State Senate District"], ["Historical Congressional District", "Historical Congressional District"], ["Native Area", "Native Area"], ["Native Statistical Area", "Native Statistical Area"], ["Native Subdivision", "Native Subdivision"], ["School District", "School District"], ["Judicial District", "Judicial District"], ["Judicial Circuit", "Judicial Circuit"], ["County Cross CD", "County Cross CD"], ["USDA County Type", "USDA County Type"], ["Hospital Referral Region", "Hospital Referral Region"], ["Hospital Service Area", "Hospital Service Area"], ["Media Market", "Media Market"], ["Continent", "Continent"], ["Country", "Country"], ["Subnational Region", "Subnational Region"], ["Urban Center", "Urban Center"], ["Country", "Country"], ["State", "State"], ["Urban Center", "Urban Center"], ["5M Person Circle", "5M Person Circle"], ["US 5M Person Circle", "US 5M Person Circle"], ["10M Person Circle", "10M Person Circle"], ["US 10M Person Circle", "US 10M Person Circle"], ["20M Person Circle", "20M Person Circle"], ["US 20M Person Circle", "US 20M Person Circle"], ["50M Person Circle", "50M Person Circle"], ["US 50M Person Circle", "US 50M Person Circle"], ["100M Person Circle", "100M Person Circle"], ["US 100M Person Circle", "US 100M Person Circle"], ["200M Person Circle", "200M Person Circle"], ["US 200M Person Circle", "US 200M Person Circle"], ["500M Person Circle", "500M Person Circle"], ["US 500M Person Circle", "US 500M Person Circle"], ["1B Person Circle", "1B Person Circle"], ["US 1B Person Circle", "US 1B Person Circle"]]
+[["State", "County"], ["Native Area", "Native Subdivision"], ["Native Statistical Area", "Native Subdivision"], ["CSA", "MSA"], ["MSA", "County"], ["County", "City"], ["CCD", "City"], ["City", "Neighborhood"], ["School District", "Neighborhood"], ["ZIP", "Neighborhood"], ["Urban Area", "City"], ["Judicial Circuit", "Judicial District"], ["County", "County"], ["MSA", "MSA"], ["CSA", "CSA"], ["Urban Area", "Urban Area"], ["ZIP", "ZIP"], ["CCD", "CCD"], ["City", "City"], ["Neighborhood", "Neighborhood"], ["Congressional District", "Congressional District"], ["State House District", "State House District"], ["State Senate District", "State Senate District"], ["Historical Congressional District", "Historical Congressional District"], ["Native Area", "Native Area"], ["Native Statistical Area", "Native Statistical Area"], ["Native Subdivision", "Native Subdivision"], ["School District", "School District"], ["Judicial District", "Judicial District"], ["Judicial Circuit", "Judicial Circuit"], ["County Cross CD", "County Cross CD"], ["USDA County Type", "USDA County Type"], ["Hospital Referral Region", "Hospital Referral Region"], ["Hospital Service Area", "Hospital Service Area"], ["Media Market", "Media Market"], ["Continent", "Continent"], ["Country", "Country"], ["Subnational Region", "Subnational Region"], ["Urban Center", "Urban Center"], ["5M Person Circle", "5M Person Circle"], ["10M Person Circle", "10M Person Circle"], ["20M Person Circle", "20M Person Circle"], ["50M Person Circle", "50M Person Circle"], ["100M Person Circle", "100M Person Circle"], ["200M Person Circle", "200M Person Circle"], ["500M Person Circle", "500M Person Circle"], ["1B Person Circle", "1B Person Circle"], ["Country", "Country"], ["State", "State"], ["Urban Center", "Urban Center"], ["US 5M Person Circle", "US 5M Person Circle"], ["US 10M Person Circle", "US 10M Person Circle"], ["US 20M Person Circle", "US 20M Person Circle"], ["US 50M Person Circle", "US 50M Person Circle"], ["US 100M Person Circle", "US 100M Person Circle"], ["US 200M Person Circle", "US 200M Person Circle"], ["US 500M Person Circle", "US 500M Person Circle"], ["US 1B Person Circle", "US 1B Person Circle"]]
