+++ urbanstats/consolidated_data/produce_consolidated_data.py
@@ -5,7 +5,7 @@ import shapely.geometry
 import tqdm.auto as tqdm
 
 from urbanstats.website_data.output_geometry import convert_to_protobuf
-from shapefiles import filter_table_for_type, load_file_for_type, shapefiles
+from urbanstats.geometry.shapefiles.shapefiles_list import filter_table_for_type, load_file_for_type, shapefiles
 from urbanstats.protobuf import data_files_pb2
 from urbanstats.protobuf.utils import write_gzip
