+++ urbanstats/geometry/shapefiles/utils.py
@@ -1,0 +1,12 @@
+import us
+
+
+def abbr_to_state(x):
+    if "-" in x:
+        return "-".join(abbr_to_state(t) for t in x.split("-"))
+    return us.states.lookup(x).name
+
+
+def name_components(x, row, abbreviate=False):
+    name, state = row.NAME.split(", ")
+    return (name + " " + x, (state if abbreviate else abbr_to_state(state)), "USA")
