+++ urbanstats/scripts/compute_gpw_data.py
@@ -1,4 +1,4 @@
-from shapefiles import shapefiles
+from urbanstats.geometry.shapefiles.shapefiles_list import shapefiles
 from urbanstats.data.gpw import compute_gpw_data_for_shapefile_table
 
 ys = [y for x, y in shapefiles.items() if y.include_in_gpw]
