+++ urbanstats/weather/era5.py
@@ -248,7 +248,7 @@ def precipitation_statistics(bounding_box, year, month):
 
 @permacache("urbanstats/weather/era5/bounding_boxes")
 def bounding_boxes():
-    from shapefiles import shapefiles
+    from urbanstats.geometry.shapefiles.shapefiles_list import shapefiles
 
     shape = shapefiles["states"].load_file().copy()
     shape["state"] = shape.shortname.apply(lambda x: us.states.lookup(x))
